# api-contract/openapi.yaml
openapi: 3.0.3
info:
  title: Communication_LTD Ticket System API
  description: |
    API for Communication_LTD customer support ticket system
    
    ## Logging Requirements
    
    - All successful mutating operations (POST, PUT, PATCH, DELETE) with valid CSRF tokens are logged at INFO level
    - Missing or invalid CSRF tokens are logged at WARNING level
    - Requests approaching rate limit thresholds are logged at WARNING level
    - All requests include client IP, user ID (if authenticated), timestamp, and operation details
    
    ## Rate Limiting
    
    The API implements rate limiting to protect against abuse:
    - Authentication endpoints: 10 requests per minute per IP
    - Standard endpoints: 60 requests per minute per authenticated user
    - Admin endpoints: 30 requests per minute per admin user
    
    When requests approach 80% of their limit threshold, WARNING level logs are generated
  version: 1.0.0
  contact:
    name: Communication_LTD Support
servers:
  - url: /api/v1
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    CSRFToken:
      name: X-CSRF-Token
      in: header
      description: >
        CSRF token for mutation requests protection.
        All POST, PUT, PATCH, DELETE operations require this header.
        Missing or invalid tokens will be logged at WARNING level.
        Valid tokens on successful mutating requests will be logged at INFO level.
      required: true
      schema:
        type: string
...
  schemas:
    Error:
      type: object
      properties:
        code:
          type: integer
          description: HTTP status code
        message:
          type: string
          description: Error message
        details:
          type: object
          description: Additional error details
      required:
        - code
        - message
      example:
        code: 400
        message: Bad Request
        details:
          field: "username"
          reason: "Username is already taken"
    User:
      type: object
      properties:
        id:
          type: integer
          format: int64
        username:
          type: string
        email:
          type: string
          format: email
        role:
          type: string
          enum: [admin, support, customer]
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        last_login:
          type: string
          format: date-time
          nullable: true
        account_locked:
          type: boolean
          default: false
      required:
        - id
        - username
        - email
        - role
        - created_at
        - updated_at
        - account_locked
      example:
        id: 1
        username: "john_doe"
        email: "john.doe@example.com"
        role: "customer"
        created_at: "2025-08-01T10:00:00Z"
        updated_at: "2025-08-01T10:00:00Z"
        last_login: "2025-08-10T15:30:00Z"
        account_locked: false
    UserCreate:
      type: object
      properties:
        username:
          type: string
          minLength: 4
          maxLength: 50
        email:
          type: string
          format: email
        password:
          type: string
          format: password
          minLength: 10
        role:
          type: string
          enum: [admin, support, customer]
      required:
        - username
        - email
        - password
        - role
      example:
        username: "john_doe"
        email: "john.doe@example.com"
        password: "SecureP@ssw0rd123"
        role: "customer"
    UserLogin:
      type: object
      properties:
        username_or_email:
          type: string
        password:
          type: string
          format: password
      required:
        - username_or_email
        - password
      example:
        username_or_email: "john.doe@example.com"
        password: "SecureP@ssw0rd123"
    TokenResponse:
      type: object
      properties:
        access_token:
          type: string
        token_type:
          type: string
          example: "bearer"
        expires_in:
          type: integer
          description: Token expiration time in seconds
      required:
        - access_token
        - token_type
        - expires_in
      example:
        access_token: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
        token_type: "bearer"
        expires_in: 3600
...
paths:
  /auth/register:
    post:
      summary: Register a new user
      operationId: registerUser
      tags:
        - Authentication
      parameters:
        - $ref: '#/components/parameters/CSRFToken'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserCreate'
      responses:
        '201':
          description: User created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description: Conflict - Username or email already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /auth/login:
    post:
      summary: Login to get access token
      operationId: loginUser
      tags:
        - Authentication
      parameters:
        - $ref: '#/components/parameters/CSRFToken'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserLogin'
      responses:
        '200':
          description: Successful login
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenResponse'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Account locked
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /auth/me:
    get:
      summary: Get current user information
      operationId: getCurrentUser
      tags:
        - Authentication
      security:
        - bearerAuth: []
      responses:
        '200':
          description: User information
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...